+++ src/app/home/home.page.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { StorageService } from '../storage.service';
+import { SubscriptionService } from '../services/subscription.service'; 
 
 @Component({
   selector: 'app-home',
@@ -10,9 +11,13 @@ import { StorageService } from '../storage.service';
 export class HomePage implements OnInit {
 public loggedin: boolean;
   constructor(
+    private subService : SubscriptionService,
     public router: Router,
     private storage: StorageService
-  ) { }
+  ) { 
+    this.subService.GetIAPSubscription();
+
+  }
 
   async ngOnInit() {
     const profile = await this.storage.GetCurrentUserProfile();
