+++ src/app/subscribe/subscribe.page.html
@@ -11,39 +11,65 @@
   </ion-toolbar>
 </ion-header>
 <ion-content>
+
+  <ion-row>
+    <ion-col>
+      <div *ngFor="let p of products" detail="false">
+        <p> {{ p.title }}</p>
+        <p>{{ p.description }}</p>
+        <p>{{ p.price }} {{ p.currency }}</p>
+        <ion-button *ngIf="!p.owned; else owned" (click)="purchase(p)">{{'Subscribe Now' | translate}}
+        </ion-button>
+        <ng-template #owned>
+          <p>You Own This Subscription</p>
+          <p>{{'Your subscription is valid until and will renew on' | translate}} {{p.expiryDate}} </p>
+        </ng-template>
+        <hr />
+      </div>
+    </ion-col>
+  </ion-row>
+  <ion-button expand="full" (click)="restore()">Restore</ion-button>'
+  <P>To cancel any subscriptions you may go to Play Store to the Payments And Subscriptions section</P>
+</ion-content>
+
+
+<!-- <ion-content>
   <ion-row *ngIf="subscriptionEndDate">
     <ion-col *ngIf="autoRenewal; else validuntil">
-      {{'Your subscription is valid until and will renew on' | translate}} {{subscriptionEndDate}}  <ion-button (click)="CancelRenewal()">  {{'Cancel Automatic Renewal' | translate}}</ion-button>
+      {{'Your subscription is valid until and will renew on' | translate}} {{subscriptionEndDate}} <ion-button
+        (click)="CancelRenewal()"> {{'Cancel Automatic Renewal' | translate}}</ion-button>
     </ion-col>
     <ng-template #validuntil>
       <ion-col>
         {{'Your subscription is valid until' | translate}} {{subscriptionEndDate}}
       </ion-col>
     </ng-template>
-    
+
   </ion-row>
   <ion-row *ngIf="!subscriptionEndDate">
     <ion-col style="padding-top: 5%; color: red;">
       {{'Your free trial or subscription has ended' | translate}}
-    </ion-col> 
-    
-  </ion-row>
+    </ion-col>
 
-  <ion-row>
+  </ion-row> -->
+
+<!-- <ion-row>
     <ion-col>
       <p *ngIf="!subscriptionEndDate">1 {{'Month FREE' | translate}}</p>
       <p>+ {{'$5/month for 11 months' | translate}}</p>
-      <ion-button [disabled]="buttonDisabled" (click)="SubscribeYearly()">{{'Subscribe Yearly' | translate}}</ion-button>
-    <hr/>
-     
+      <ion-button [disabled]="buttonDisabled" (click)="SubscribeYearly()">{{'Subscribe Yearly' | translate}}
+      </ion-button>
+      <hr />
+
       <p> {{'$5/month' | translate}}</p>
-      <ion-button  [disabled]="buttonDisabled" (click)="SubscribeMonthly()">{{'Subscribe Monthly' | translate}}</ion-button>
+      <ion-button [disabled]="buttonDisabled" (click)="SubscribeMonthly()">{{'Subscribe Monthly' | translate}}
+      </ion-button>
     </ion-col>
-    
-  </ion-row>
-  <!-- <ion-row>
+
+  </ion-row> -->
+<!-- <ion-row>
       <ion-col>
         <ion-button (click)="cancelSubscription()">{{'Cancel Subscription' | translate}}</ion-button>
       </ion-col>    
   </ion-row> -->
-</ion-content>
+<!-- </ion-content> -->
